modularize Search component
